Now Working on OOP Style
Long time no see, I had been not able to use my PC for a long time. So was not able to code with this project. However, I am now able to access my PC with all devices, so I would be updating this project oftenly.

The main goal of this branch is to provide
- PIP simple installation
- OOP concept (so that others can contribute)

Sorry for late commits and update.
Thanks.